project/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
redis==5.0.8
psycopg2-binary==2.9.9
prometheus-client==0.21.0
python-multipart==0.0.12
